requirements.txt
Flask==1.1.2
numpy==1.19.4
pyyaml==5.3.1

# The following are needed for the structure importer
ase==3.20.1
qe-tools==1.1.4
pymatgen==2020.10.20